Extracted error handling; improved config file handling; added docs.
This app will now try to reconnect to a bridge (either initially, or after a connection drop) forever.
Added the ability to pass in a custom PHSDKListener implementation to the HueBridgeService.
About to abstract the entire Hue connection logic to another project all together.

+++ src/main/groovy/te/light_jockey/core/ConfigHandler.groovy
@@ -21,26 +21,23 @@ class ConfigHandler {
 
         temporaryDirectory = new File(tempDirPath)
         configurationFile = new File(temporaryDirectory, CONFIG_FILE_NAME)
+
+        // Create the config file if it doesn't exist
+        if(!configurationFile.exists()) {
+            FileUtils.writeStringToFile(configurationFile, "", false)
+        }
     }
 
     void updateConfigFile(Map propsToAdd) {
         Properties configFileProps = readConfigProperties()
         configFileProps.putAll(propsToAdd)
 
-        log.debug("Updating config. file to:\n\t{}", configFileProps.collect{"$it.key: $it.value"}.join("\n\t"))
+        log.debug("Updating config file contents to:\n\t{}", configFileProps.collect{"$it.key: $it.value"}.join("\n\t"))
         configFileProps.store(new FileOutputStream(configurationFile), "LightJockey configuration file")
-    }
-
-    void createConfigFile() {
-        FileUtils.writeStringToFile(configurationFile, "", false)
     }
 
     Properties readConfigProperties() {
         loadPropertiesFromInputStream(FileUtils.openInputStream(configurationFile))
-    }
-
-    boolean configFileExists() {
-        configurationFile.exists() && configurationFile.canRead() && configurationFile.canWrite()
     }
 
     boolean configFileIsStillValid() {
